+++ src/app/features/cart/models/cart-storage.model.ts
@@ -1,0 +1,7 @@
+import { CartItem } from './cart-item';
+
+export class CartStorageModel {
+  products: CartItem[];
+  quantity: number;
+  sum: number;
+}